python-api/requirements.txt
# ═══════════════════════════════════════
# iDRAC9 WhatsApp Bot — Python Redfish API
# ═══════════════════════════════════════

# --- Core ---
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic-settings==2.*

# --- HTTP Client ---
httpx==0.28.*

# --- Retry ---
tenacity==9.*

# --- Logging ---
loguru==0.7.*

# --- Redis (optional) ---
redis==5.*
